feat: one colour per container, in every command that names it
`logs` tinted each container with a stable colour from a fixed palette;
nothing else did. The same container was magenta in `logs` and grey in
`ps`, `stats`, `images` and the `up`/`down` progress lines — so the one
thing that makes a busy project scannable worked in exactly one command.

The colour is now keyed in one place. `ui::identity_style` strips the
project name before hashing, so `proj-web-1` (what `ps` and the progress
lines print) and `web-1` (what `logs` prefixes) resolve to the same
colour. Hashing whatever string each site happened to hold was why they
disagreed; a test pins that every spelling of one container matches.

Lifecycle verbs gained bands. They were all the same green, so
`Volume data Removed` — which destroys data and cannot be undone — was
styled exactly like `Started`. Green means something now exists, yellow
that it stopped but survives, red that it is gone, dim that nothing
changed.

`stats` carried no colour at all, which is odd for the one command whose
entire purpose is finding the row in trouble: a container at 95% of its
memory limit looked exactly like one at 0.02%. CPU and memory percentages
are banded green/yellow/red at 70 and 90, and the absolute figures are
dimmed so the percentages read first.

Padding is applied before painting throughout. ANSI codes are zero-width,
so padding afterwards counts them and knocks every later column out of
alignment.

`stats.rs` reached 489 code lines against a 500 hard limit, so its tests
moved to `stats_tests.rs` — the same split `autostart` and the libpod
client already use. 337 and 153 now.

Verified against Podman 5.4.2 that all four agree:

    up     Container col-db-1[96m]   Started[green]
    ps     col-db-1[96m]   ...       running[green]
    logs   db-1 | [96m]
    down   Container col-db-1[96m]   Removed[red]
